+++ docs/changelog.rst
@@ -13,6 +13,8 @@ CHANGELOG
   stars in attachments list (fixes #1117)
 * Ability to edit attachments from detail pages directly (fixes #177, the 5th oldest issue!)
 * Add missing columns in intervention exports (fixes #1167)
+* Add ability to have paths records in database that will not appear in Geotrek
+  lists and maps. Just set column ``visible`` to false in ``l_t_troncon`` table.
 
 **Bug fixes**